move player into world crate
