--- Engine/Source/player/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "player"
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[dependencies]
-tracing.workspace = true
-
-events.workspace = true
-macros.workspace = true
-platform.workspace = true
-world.workspace = true
-
-glam.workspace = true
