+++ Engine/Source/world/Cargo.toml
@@ -13,3 +13,4 @@ tracing.workspace = true
 macros.workspace = true
 events.workspace = true
 utils.workspace = true
+platform.workspace = true
